rename to mp-mpiu
less likely to collide with names when this is split to a new package.
